+++ test_retrying.py
@@ -40,6 +40,13 @@ class TestStopConditions(unittest.TestCase):
     def test_legacy_explicit_stop_type(self):
         r = Retrying(stop="stop_after_attempt")
 
+    def test_stop_func(self):
+        r = Retrying(stop_func=lambda attempt, delay: attempt == delay)
+        self.assertFalse(r.stop(1, 3))
+        self.assertFalse(r.stop(100, 99))
+        self.assertTrue(r.stop(101, 101))
+
+
 class TestWaitConditions(unittest.TestCase):
 
     def test_no_sleep(self):
@@ -113,6 +120,13 @@ class TestWaitConditions(unittest.TestCase):
 
     def test_legacy_explicit_wait_type(self):
         r = Retrying(wait="exponential_sleep")
+
+    def test_wait_func(self):
+        r = Retrying(wait_func=lambda attempt, delay: attempt * delay)
+        self.assertEqual(r.wait(1, 5), 5)
+        self.assertEqual(r.wait(2, 11), 22)
+        self.assertEqual(r.wait(10, 100), 1000)
+
 
 class NoneReturnUntilAfterCount:
     """
@@ -282,7 +296,7 @@ class TestDecoratorWrapper(unittest.TestCase):
         self.assertTrue(t >= 250)
         self.assertTrue(result)
 
-    def test_with_stop(self):
+    def test_with_stop_on_return_value(self):
         try:
             _retryable_test_with_stop(NoneReturnUntilAfterCount(5))
             self.fail("Expected RetryError after 3 attempts")
@@ -290,6 +304,14 @@ class TestDecoratorWrapper(unittest.TestCase):
             self.assertFalse(re.last_attempt.has_exception)
             self.assertEqual(3, re.last_attempt.attempt_number)
             self.assertTrue(re.last_attempt.value is None)
+            print(re)
+
+    def test_with_stop_on_exception(self):
+        try:
+            _retryable_test_with_stop(NoIOErrorAfterCount(5))
+            self.fail("Expected IOError")
+        except IOError as re:
+            self.assertTrue(isinstance(re, IOError))
             print(re)
 
     def test_retry_if_exception_of_type(self):
@@ -303,7 +325,7 @@ class TestDecoratorWrapper(unittest.TestCase):
             print(n)
 
         try:
-            _retryable_test_with_exception_type_io_attempt_limit(NoIOErrorAfterCount(5))
+            _retryable_test_with_exception_type_io_attempt_limit_wrap(NoIOErrorAfterCount(5))
             self.fail("Expected RetryError")
         except RetryError as re:
             self.assertEqual(3, re.last_attempt.attempt_number)
@@ -323,7 +345,7 @@ class TestDecoratorWrapper(unittest.TestCase):
             print(n)
 
         try:
-            _retryable_test_with_exception_type_custom_attempt_limit(NoCustomErrorAfterCount(5))
+            _retryable_test_with_exception_type_custom_attempt_limit_wrap(NoCustomErrorAfterCount(5))
             self.fail("Expected RetryError")
         except RetryError as re:
             self.assertEqual(3, re.last_attempt.attempt_number)